backend/requirements.txt
llama-cpp-python>=0.3.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
